Add files via upload
The purpose of the lab is for students to practice implementing linear regression models with statsmodels in Python environment. The context is real estate market for a NE town called Eastville, CT. Each team is expected to conduct both descriptive analytics and predictive analytics to understand the predict price for a house. And write a short report on the findings.